requirements.txt
Django==5.1.6
python-dotenv==1.0.1
dj-database-url==2.3.0
djangorestframework-3.15.2
